chore: extract integrations methods to a file
